+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.5.1] - 2025-01-09
+
+### Fixes
+
+* Fix error in disassembly with label substitution
+* Account for label substitution in unit testing
+
 ## [3.5.0] - 2024-12-29
 
 ### Fixes